backend/requirements.txt
fastapi[all]
requests
python-dotenv
pydantic
firebase-admin
